+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   both the single and multiple element variants.
 * Export `reverseCommitsOntoChunks` from the `Unsafe` module.
 * Add `Semigroup` and `Monoid` instances for `Chunks`.
+* Add `consLengthLE32`.
 
 ## 0.3.0.0 -- 2019-10-17